:recycle: Do not prepend transport path to cache path
This allows users to specify a fully custom path for the cache, and consequently Maven artifacts to be downloaded in the .m2 directory.

Ideally, the way caches are created should be changed. Rather than having a single cache, every repository should have its cache.
Since I'm going to use this exclusively with Maven, add a TODO and improve eventually